Fix ReferenceError in handling of hotlist message
Handling the hotlist action requires a call to getState in order to get
the currentBufferId.

The nonexistent state local variable was raising a ReferenceError, so no
buffers would show as having hot messages on connect.

Also fix a ReferenceError in the hotlists reducer, which was never
reached because of the above.