Reconnect on connection loss
When WeechatRN goes into the background, at some point the WebSocket
will be closed. Track whether we've successfully connected and then
reconnect when this happens.

Unfortunately the best way to determine this seems to be via an onerror
handler, as a code is not provided for this state in onerror.