Don't set currentBufferId to a closed buffer
If weechat has restarted, or a buffer has been closed while weechatRN is
disconnected, then we will have an incorrect pointer set in
currentBufferId.

While weechat handles this just fine, weechatRN behaves like a buffer is
"mostly" open, in that the previous lines will show up but the drawer
will remain closed and the topic will be unset.

Fixes #14.