Split up workflow into multiple jobs
For finer control over each step of the publication process. The gist of
how this works is:

1. Build and publish the app via Expo
2. Download the binary from Expo
3. Upload the binary as an artifact
4. Push the artifact to Fastlane

It's a lot of uploading and downloading but because this is broken up in
multiple jobs, we don't have to risk restarting the entire process if
something fails in between. Instead, we could re-run the specific job
that we need.