Handle _buffer_cleared to fix duplicate lines
When a buffer is cleared it's possible that a pointer is reused. This
causes issues with the Buffer FlatList keyExtractor, which expects
unique values for the key.

wee-slack clears the buffer when loading a buffer or after reconnecting.
This would cause out of order messages and errors in development mode.

Fixes #12.