Forget closed buffer state when reconnecting
While the buffer list is fetched and replaced when (re)connecting,
references to closed buffers may still exist in the hotlist, nicklist,
and line states, as well as the currentBufferId field in the app state.

Ensure that references to closed buffers are cleared out from the other
states when the buffer list is fetched.

Note that there is still a potential issue during upgrades if a pointer
happens to be reused. Checking whether an upgade occurred while offline
would be ideal, but it's an edge case.