Test connection and handle authentication failure
Weechat indicates authentication failure by closing the socket.
Previously, this would result in onerror being called, causing a
reconnect loop. onerror no longer seems to be called when the connection
is closed, resolving the original issue. However, in order to present an
error to the user on authentication falure, we need to detect if onclose
was called during authentication.

Fixes #30