Implement media upload from library or camera
Only unauthenticated or basic auth protected endpoints are supported at
this time. Uploads via POST with the image in the form field `file`.
The response body is pasted to the input field.

This required introducing a new SettingsNavigator component so as to not
overload the LoginForm. In the future, we'll bring users straight to the
buffer list (and attempt to connect), so the LoginForm will not be the
home screen.

Closes #62