Remove unregistered push tokens
Per Expo documentation, when the API responds with DeviceNotRegistered,
notifications must not be sent to this device any further.

Unfortuntaely this has proven difficult to test. After 24 hours, tokens
from an uninstalled device and device with disabled notifications are
still valid. However, testing with a completely invalid token (e.g.
"invalid") shows that this works as expected.