Open buffer from push notification
When launched by a push notification response, the connection may be in
a known or not yet known (onerror not yet fired) disconnected state. In
order to bring the user to the buffer, the buffer list needs to be up to
date, which requires known connection state.

While the connection state is correct in the onResume handler, there is
no AppState change for the initial render, and componentDidMount fires
before the notification listener is attached. (Things might have been
easier if this behaved like in Android onResume.)

To work around this, the notification response handler dispatches a new
action which wraps a bufferNotificationAction. This wrapper action is
picked up by a listener that requests the buffer list and awaits the
response.

If connected, the buffer list response will cause the listener to
dispatch the wrapped bufferNotificationAction. If disconnected (or
onerror not yet fired), the listener will block until connection is
re-established, before dispatching the bufferNotificationAction.