Don't unnecessarily fetch lines for notification
When a notification is received for the current buffer and we're still
connected (but in the background), new lines will already have loaded
when the app resumes However, the lines may not have rendered yet, so
they cannot be scrolled to reliably.

Work around this by changing the FlatList key, resetting its internal
measurement state. Subsequent scroll will then either make it to the
target or cycle trigger onScrollToIndexFailed.