Migrate back to FlashList
Scroll to the target index in a loop until the layout returned by the
layout manager isOverridden. This indicates that the target cell has
actually been measured and can be scrolled to reliably.

Similarly to FlatList, we must reset the internal measurement state via
key prop to work around the layout manager indicating that the target
index has been measured. This is because the prepended items may have
invalidated the isOverridden indication for the given index.