Use stable IDs to dispatch notifications
Dispatching notifications using pointers is error prone, especially with
wee-slack, which is very likely to reuse pointers when it reconnects.
Stable IDs were introduced in 4.3.0 for buffers and 4.4.0 for lines, and
remain stable throughout the lifetime of the weechat process (including
upgrades.)

The notification script has been updated to use stable IDs when
supported. We fallback to the buffer pointer or line pointer as an
integer when stable IDs are not available.