Restore content offset after background snapshot
In iOS 26, backgrounding an app triggers an app snapshot [1], which will
be taken in all supported device orientations. When in landscape mode,
the OS will make a layout pass in portrait followed by another in
landscape (and vice versa), taking a snapshot of each.

If scrolled to the bottom of a ScrollView in landscape mode, the content
offset will be incorrect on resume, as UIScrollView updates the offset
during the portrait layout pass to prevent overscroll. When the next
layout pass is made, the offset is not reset to the original value.

Store the content offset on update when foregrounded. When backgrounded,
override the content offset with the stored content offset, clamped to
the content size. When the second snapshot is taken (in the original
orientation), the original content offset is restored.

[1] https://developer.apple.com/documentation/uikit/preparing-your-ui-to-run-in-the-background#Prepare-your-UI-for-the-app-snapshot